Migrated TrashSimple to ts
